firmware: cirrus: arm64 compiler fix
When building for a 64-bit system, the compiler
will raise an error since we are attempting to
take the maximum value of a 64-bit size-t and a
32-bit int.

+++ drivers/firmware/cirrus/cl_dsp.c
@@ -236,7 +236,7 @@ static int cl_dsp_read_wt(struct cl_dsp *dsp, int pos, int size)
 		if (entry->type == WT_TYPE_TERMINATOR) {
 			dsp->wt_desc->owt.nwaves = i;
 			dsp->wt_desc->owt.bytes = max(ch.bytes,
-					(void *)max - buf);
+					(int)((void*)max - buf));
 
 			return dsp->wt_desc->owt.bytes;
 		}
